Extract LoadDumpPackManifest to extension class
Moves the LoadDumpPackManifest logic from a private static method in FrameworkDefinition.All.cs into a new public extension method in FrameworkDefinitionExtension.cs. Updates CreateAll() to call the extension method on the frameworks list.
